requirements.txt
Django==4.2
psycopg2-binary
asgiref==3.8.1
django-filter==23.1  
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
six==1.16.0
pytest
pytest-django
pytest-cov
pytest-docker
mixer
django-cors-headers == 4.4.0
dj_rest_auth==6.0.0